refactor : Column 어노테이션 수정, 및 Dto 수정
