--- com.sparta.logistics.product/src/main/java/com/sparta/logistics/product/application/dto/StockRequestDto.java
@@ -1,12 +1,0 @@
-package com.sparta.logistics.product.application.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class StockRequestDto {
-    private int orderQuantity; // feign client dto
-}
